Struct soft_ascii_string::SoftAsciiStr

#[repr(transparent)]
pub struct SoftAsciiStr(_);

A str wrapper adding a "is us-ascii" soft constraint.

This means that it should be ascii but is not guaranteed to be ascii. Which means non ascii chars are not a safety issue just a potential bug.

This is useful for situations where:

you would have many unsafe from str conversions/"unnecessary" checks with a strict AsciiStr

you rarely have to strictly rely on the value being ascii

Note

Some functions which should be implemented directly on SoftAsciiStr like e.g. trim_matches are only provided through .as_str(). This is because the Pattern API and SliceIndex API is unstable i.e. can only be implemented in unstable for now. Once it gets stabilized (rust #27721/#35729) implementations can be added

Methods

impl SoftAsciiStr

pub fn from_unchecked(s: &str) -> &SoftAsciiStr

pub fn from_str_unchecked(s: &str) -> &SoftAsciiStr

Deprecated since 1.0.0:
use from_unchecked

pub fn from_unchecked_mut(s: &mut str) -> &mut SoftAsciiStr

pub fn from_str(source: &str) -> Result<&Self, FromSourceError<&str>>

pub fn revalidate_soft_constraint(&self) -> Result<&Self, FromSourceError<&str>>

reruns checks if the "is us-ascii" soft constraint is still valid

pub fn as_str(&self) -> &str

pub fn into_soft_ascii_string(self: Box<SoftAsciiStr>) -> SoftAsciiString

pub fn from_boxed_str(bs: Box<str>) -> Box<SoftAsciiStr>

pub fn into_boxed_str(self: Box<SoftAsciiStr>) -> Box<str>

pub fn lines(&self) -> SoftAsciiLines

pub fn split_whitespace(&self) -> SoftAsciiSplitWhitespace

pub fn char_indices(&self) -> SoftAsciiCharIndices

pub fn chars(&self) -> SoftAsciiChars

pub fn split_at(&self, mid: usize) -> (&SoftAsciiStr, &SoftAsciiStr)

pub unsafe fn slice_unchecked(&self, begin: usize, end: usize) -> &SoftAsciiStr

Deprecated since 1.1.0:
deprecated in std

pub unsafe fn get_unchecked<I>(&self, index: I) -> &SoftAsciiStr where
I: TempSliceIndexHelper,

Proxy of std::str::get_unchecked.

Currently limited to the various range types:

Range<usize>

RangeInclusive<usize>

RangeFrom<usize>

RangeTo<usize>

RangeToInclusive<usize>

RangeFull

Once all methods on SliceIndex are stable this can be implemented using SliceIndex<SoftAsciiStr> bounds.

pub fn inner_str_mut(&mut self) -> &mut str

returns a mutable str reference to the inner buffer

Soft Constraint

be aware that it is very easy to introduce bugs when directly editing a SoftAsciiStr as an str. Still compared to a AsciiStr implementation this won't introduce unsafety, just possible brakeage of the soft constraint that the data should be ascii.

pub fn parse<F>(&self) -> Result<F, <F as FromStr>::Err> where
F: FromStr,

impl SoftAsciiStr

pub fn to_lowercase(&self) -> SoftAsciiString

impl SoftAsciiStr

pub fn to_uppercase(&self) -> SoftAsciiString

impl SoftAsciiStr

pub fn repeat(&self, n: usize) -> SoftAsciiString

impl SoftAsciiStr

pub fn trim_right(&self) -> &SoftAsciiStr

Deprecated since 1.1.0:
deprecated in std

pub fn trim_left(&self) -> &SoftAsciiStr

Deprecated since 1.1.0:
deprecated in std

pub fn trim_end(&self) -> &SoftAsciiStr

pub fn trim_start(&self) -> &SoftAsciiStr

pub fn trim(&self) -> &SoftAsciiStr

impl SoftAsciiStr

pub fn len(&self) -> usize

pub fn is_empty(&self) -> bool

pub fn is_char_boundary(&self, index: usize) -> bool

pub fn as_ptr(&self) -> *const u8

pub fn encode_utf16(&self) -> EncodeUtf16

pub fn is_ascii(&self) -> bool

pub fn as_bytes(&self) -> &[u8]